refactor into FeedTab
